Workaround for a suppositional bug in pgpro_scheduler: a worker of a main scheduled job does not see his submitted one-time jobs in case when all "at job executors" are busy (when there is no any active other jobs, one-time subjobs filtered by current executor are seen)

+++ etl-engine/etl-engine-intra-database/src/main/procedures/p_wait_for_scheduled_task_subjobs_completion.sql
@@ -75,7 +75,7 @@ begin
 			exit;
 		end if;
 		
-		if extract(hours from clock_timestamp() - l_start_timestamp) >= i_timeout_in_hours then
+		if extract(hours from clock_timestamp() - l_start_timestamp)::integer >= i_timeout_in_hours then
 			raise exception 'Timeout occured while waiting for the scheduled task completion: %', i_scheduled_task_name;
 		end if;
